# `ls`: file name handling performance improvements