I did some more performance improvements based on flamegraph data. I documented the method for this in `BENCHMARKING.md`. The changes I made are mostly reducing some allocations and reducing the times that `OsStrs` are converted to `Strings`.

**Before**
```
Benchmark #1: ls ../firefox-source/ -R
  Time (mean ± σ):     741.3 ms ±  26.7 ms    [User: 521.4 ms, System: 216.2 ms]
  Range (min … max):   708.8 ms … 766.3 ms    10 runs
 
Benchmark #2: target/release/coreutils ls ../firefox-source/ -R
  Time (mean ± σ):      1.049 s ±  0.031 s    [User: 831.8 ms, System: 212.2 ms]
  Range (min … max):    1.005 s …  1.090 s    10 runs
 
Summary
  'ls ../firefox-source/ -R' ran
    1.41 ± 0.07 times faster than 'target/release/coreutils ls ../firefox-source/ -R'
```

**After**
```
Benchmark #1: ls ../firefox-source/ -R
  Time (mean ± σ):     542.1 ms ±   3.4 ms    [User: 375.2 ms, System: 164.6 ms]
  Range (min … max):   537.9 ms … 547.2 ms    10 runs
 
Benchmark #2: target/release/coreutils ls ../firefox-source/ -R
  Time (mean ± σ):     649.3 ms ±   9.3 ms    [User: 479.6 ms, System: 167.4 ms]
  Range (min … max):   638.3 ms … 664.2 ms    10 runs
 
Summary
  'ls ../firefox-source/ -R' ran
    1.20 ± 0.02 times faster than 'target/release/coreutils ls ../firefox-source/ -R'
```